Encryption of the text
This code encrypts the message by getting the word-by-word arrangement of the variable 'key' to the input text. To enumerate,
suppose that 123 is the list of the variable 'chars' and the shuffled variable 'key' is abc. When the input typed by the user is '123' then the encrypted text will be abc.